Refactored featured sites folder
