--- app/brochure/routes/featured/review.js
@@ -1,1 +1,0 @@
-review.js
