--- app/brochure/routes/featured/sites.json
@@ -1,81 +1,0 @@
-[
-  {
-    "name": "Rebecca Kukshtel",
-    "image": "rebecca.jpg",
-    "domain": "rebeccatron.com",
-    "url": "https://www.rebeccatron.com",
-    "description": "is an engineer based in New York City. Rebecca",
-    "template": "Rosa",
-    "first": true
-  },
-  {
-    "name": "Rachel Pietraszek",
-    "image": "rachel.jpg",
-    "domain": "rachelpietraszek.com",
-    "url": "https://photos.rachelpietraszek.com",
-    "description": "is a lawyer based in Toronto. Rachel",
-    "template": "Photos"
-  },
-  {
-    "name": "Andrew Janjigian",
-    "image": "andrew.jpg",
-    "domain": "andrewjanjigian.com",
-    "url": "https://blog.andrewjanjigian.com",
-    "description": "is a photographer and editor at Cook’s Illustrated. Andrew",
-    "template": "Rosa"
-  },
-
-  {
-    "name": "Aurynn Shaw",
-    "image": "aurynn.jpg",
-    "domain": "blog.aurynn.com",
-    "url": "https://blog.aurynn.com",
-    "description": "is a writer and engineer from New Zealand. Aurynn’s site",
-    "custom": true
-  },
-
-  {
-    "name": "John Beeler",
-    "image": "beeler.jpg",
-    "domain": "johnbeeler.com",
-    "url": "https://com.johnbeeler.com",
-    "description": "is the label director at Asthmatic Kitty Records. John",
-    "template": "Default"
-  },
-
-  {
-    "name": "Hanna Utkin",
-    "image": "hanna.jpg",
-    "domain": "hannautkin.com",
-    "url": "https://hannautkin.com",
-    "description": "is a filmmaker and producer from New York City. Hanna",
-    "template": "Console"
-  },
-
-  {
-    "name": "Matthew Battles",
-    "image": "matthew.jpg",
-    "domain": "matthewbattles.org",
-    "url": "https://www.matthewbattles.org",
-    "description": "directs metaLAB at Harvard. Matthew’s website",
-    "template": "Rosa"
-  },
-
-  {
-    "name": "John Pavlus",
-    "image": "john.jpg",
-    "domain": "pavlusoffice.com",
-    "url": "https://john.pavlusoffice.com",
-    "description": "is a writer and filmmaker based in Portland, Oregon. John",
-    "custom": true
-  },
-
-  {
-    "name": "Jacoby Young",
-    "image": "jacoby.png",
-    "domain": "jacobyyoung.com",
-    "url": "https://www.jacobyyoung.com/about",
-    "description": "works at Kapālama Elementary School in Hawaii. Jacoby",
-    "custom": true
-  }
-]
