+++ app/scheduler.js
@@ -1,6 +1,7 @@
 var Entries = require("./models/entries");
 var Entry = require("./models/entry");
 var User = require("./models/user");
+var updateFeaturedSites = require('./brochure/routes/featured/update');
 var async = require("async");
 var schedule = require("node-schedule").scheduleJob;
 var Blog = require("./models/blog");
@@ -72,6 +73,13 @@ module.exports = function() {
     // Start the backup daemon
     console.log("Backup: It is 1am, time to start!");
     backup.now();
+  });
+
+  console.log("Scheduled daily check of the featured sites for midnight!");
+  schedule({ hour: 8, minute: 0 }, function() {
+    // Start the backup daemon
+    console.log("Backup: It is 1am, time to start!");
+    updateFeaturedSites();
   });
 
   // At some point I should check this doesnt consume too
